Move helpers to a separate file
